+++ CHANGES.rst
@@ -10,6 +10,12 @@ Unreleased
 * :meth:`.submit`, :meth:`.submit_image`, and :meth:`.submit_video` can be used
   to submit a post directly to a collection.
 
+**Changed**
+
+* :meth:`.Reddit.info` now accepts any non-str iterable for fullnames.
+* :meth:`.Reddit.info` now returns a generator instead of a list when
+  using the ``url`` parameter.
+
 6.2.0 (2019/05/05)
 ------------------